requirements.txt
flask
torchvision
pillow